+++ internal/api/web/_app/immutable/entry/start.DWjeecp-.js
@@ -1,0 +1,1 @@
+import{o as e,t}from"../chunks/CFFv0j_P.js";export{e as load_css,t as start};